Editing, deleting comments
Added notes which are stored as files and the ability to post a note
as a comment. Notes are stored in a directory. The editor used is
based on the EDITOR environment variable, or vi, or ed. There is a new
command to edit notes and to list notes, and to set the editor.

The pager defaults to the PAGER environment variable.

Removed the rc path ~/.config/.jan-pona-mute; added
~/.jan-pona-mute.d/login instead.

Fixed display of index for comments. It is now possible to use this
index to delete comments. The delete command also deletes notes.